+++ mvs_eland_tool/mvs_eland_tool.py
@@ -43,7 +43,7 @@ import src.F0_output as output_processing
 def main(**kwargs):
     # Display welcome text
     version = (
-        "0.0.2"  # update_me Versioning scheme: Major release.Minor release.Patches
+        "0.1.1"  # update_me Versioning scheme: Major release.Minor release.Patches
     )
     date = "25.11.2019"  # update_me Update date
 
@@ -78,12 +78,17 @@ def main(**kwargs):
     if not user_input["path_input_file"].endswith("json"):
         logging.debug("Accessing script: A1_csv_to_json")
         path_to_json_from_csv = load_data_from_csv.create_input_json(
-            input_directory=user_input["path_input_folder"]
+            input_directory=user_input["path_input_file"]
         )
         user_input.update({"path_input_file": path_to_json_from_csv})
 
     logging.debug("Accessing script: B0_data_input_json")
-    dict_values = data_input.load_json(user_input["path_input_file"])
+    dict_values = data_input.load_json(
+        user_input["path_input_file"],
+        path_input_folder=user_input["path_input_folder"],
+        path_output_folder=user_input["path_output_folder"],
+        path_input_sequences=user_input["path_input_sequences"],
+    )
 
     print("")
     logging.debug("Accessing script: C0_data_processing")