+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Here is a template for new release sections
 
 ### Fixed
 - Input directory of csv files specified by user is handed to `load_data_from_csv.create_input_json()` (#112)
+- \#111 & \#114 fix user choice of output folder via command line arguments(#115)
 
 ## [0.1.0] -2020-01-30